Create 06 - Whitelist.sol
A whitelist is a security list that provides access to only pre-approved addresses. Whatever is on the "list" gets access to resources, whereas the rest are denied access.

Whitelisting is the opposite of blacklisting, as you might expect. Blacklisting blocks specific sites, services, or apps, whereas whitelisting uses specifics to place more control in the hands of network administrators.